Fixed a bug for showing movie json after deletion of all genres from a movie

+++ service/movie_service.go
@@ -5,6 +5,7 @@ import (
 	m "movies_api/models"
 	r "movies_api/repository"
 	"slices"
+	"strings"
 )
 
 type MovieService struct {
@@ -12,7 +13,7 @@ type MovieService struct {
 }
 
 func (s *MovieService) GetAllMovies() ([]m.MovieDto, error) {
-	movies, err := s.repo.FindAllMovies()
+	movies, err := s.repo.FindAllMovies(false, "")
 	if err != nil {
 		return []m.MovieDto{}, err
 	}
@@ -44,7 +45,7 @@ func (s *MovieService) GetAllMoviesWithGenre(genreID int) ([]m.MovieDto, error) 
 
 func (s *MovieService) GetAllMoviesWithYear(year int) ([]m.Movie, error) {
 	if year < 1885 || year > 2050 || year == 0 {
-		return []m.Movie{}, fmt.Errorf("%w Invalid release year.", m.ErrInvalidInput)
+		return []m.Movie{}, fmt.Errorf("%w: invalid release year.", m.ErrInvalidInput)
 	}
 
 	movies, err := s.repo.FindMoviesByYear(year)
@@ -56,6 +57,19 @@ func (s *MovieService) GetAllMoviesWithYear(year int) ([]m.Movie, error) {
 
 func (s *MovieService) GetAllMoviesWithActor(actorID int) ([]m.MovieDto, error) {
 	movies, err := s.repo.FindMoviesWithActor(actorID)
+	if err != nil {
+		return []m.MovieDto{}, err
+	}
+
+	sortMovies(movies)
+
+	return movies, nil
+}
+
+func (s *MovieService) GetAllMoviesWithTitle(title string) ([]m.MovieDto, error) {
+	title = strings.ToLower(title)
+
+	movies, err := s.repo.FindAllMovies(true, title)
 	if err != nil {
 		return []m.MovieDto{}, err
 	}
@@ -111,9 +125,13 @@ func (s *MovieService) MoviePatcher(movieData m.MoviePatchDto, movieID int) (m.M
 	return movie, nil
 }
 
-func (s *MovieService) DeleteMovie(movieID int) (bool, error) {
+func (s *MovieService) DeleteMovie(movieID int, force bool) (bool, error) {
 	if movieID <= 0 {
-		return false, fmt.Errorf("%w invalid movie id.", m.ErrBadRequest)
+		return false, fmt.Errorf("%w: invalid movie id.", m.ErrBadRequest)
+	}
+
+	if !force {
+		return false, fmt.Errorf("%w: cannot delete movie, use \"force\" to perform deletion.", m.ErrBadRequest)
 	}
 
 	ok, err := s.repo.DeleteMovieByID(movieID)
@@ -135,11 +153,11 @@ func (s *MovieService) FindActorsInMovie(movieID int) ([]m.ActorInFilmDto, error
 
 func (s *MovieService) AddActorToMovie(movieID, actorID int) (m.MovieDto, error) {
 	if movieID <= 0 {
-		return m.MovieDto{}, fmt.Errorf("%w invalid movie id.", m.ErrBadRequest)
+		return m.MovieDto{}, fmt.Errorf("%w: invalid movie id.", m.ErrBadRequest)
 	}
 
 	if actorID <= 0 {
-		return m.MovieDto{}, fmt.Errorf("%w invalid actor id.", m.ErrBadRequest)
+		return m.MovieDto{}, fmt.Errorf("%w: invalid actor id.", m.ErrBadRequest)
 	}
 
 	err := s.repo.AddActorToMovie(movieID, actorID)
@@ -157,19 +175,19 @@ func (s *MovieService) AddActorToMovie(movieID, actorID int) (m.MovieDto, error)
 
 func (s *MovieService) DeleteActorFromMovie(movieID, actorID int) (m.MovieDto, error) {
 	if movieID <= 0 {
-		return m.MovieDto{}, fmt.Errorf("%w invalid movie id.", m.ErrBadRequest)
+		return m.MovieDto{}, fmt.Errorf("%w: invalid movie id.", m.ErrBadRequest)
 	}
 
 	if actorID <= 0 {
-		return m.MovieDto{}, fmt.Errorf("%w invalid actor id.", m.ErrBadRequest)
+		return m.MovieDto{}, fmt.Errorf("%w: invalid actor id.", m.ErrBadRequest)
 	}
 
-	err := s.repo.RemoveActorFromMovie(movieID, actorID)
+	movie, err := s.repo.FindMovieByID(movieID)
 	if err != nil {
 		return m.MovieDto{}, err
 	}
 
-	movie, err := s.repo.FindMovieByID(movieID)
+	err = s.repo.RemoveActorFromMovie(movieID, actorID)
 	if err != nil {
 		return m.MovieDto{}, err
 	}
